Consolidate pricing models
